[REFACT] 기존 calendar 리덕스를 modalEvents로 변경하기 #NADA-50
