+++ src/modules/index.js
@@ -4,13 +4,13 @@ import activity from './activity';
 import barStatus from './bar';
 import schedule from './schedule';
 import user from './user';
-import calendar from './calendar/calendar';
+import modalEvent from './calendar/modalEvent';
 
 const rootReducer = combineReducers({
   auth,
   schedule,
   user,
-  calendar,
+  modalEvent,
   activity,
   barStatus,
 });
